+++ src/main/kotlin/com/github/qlonik/intellij/peggy/PeggyIcons.kt
@@ -1,0 +1,7 @@
+package com.github.qlonik.intellij.peggy
+
+import com.intellij.openapi.util.IconLoader
+
+object PeggyIcons {
+    val FILETYPE = IconLoader.getIcon("/icons/PeggyFile.svg")
+}