+++ src/main/kotlin/com/github/qlonik/intellij/peggy/PeggyLanguage.kt
@@ -1,0 +1,9 @@
+package com.github.qlonik.intellij.peggy
+
+import com.intellij.lang.Language
+
+class PeggyLanguage private constructor() : Language("Peggy") {
+    companion object {
+        val INSTANCE = PeggyLanguage()
+    }
+}