+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ## [Unreleased]
 ### Added
+- Initial definition of Peggy file type
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)